Tests: Tidy up compiled tests
