+++ main/main.h
@@ -73,10 +73,8 @@ public:
 	static Error setup2(bool p_show_boot_logo = true); // The thread calling setup2() will effectively become the main thread.
 	static String get_locale_override();
 	static void setup_boot_logo();
-#ifdef TESTS_ENABLED
 	static Error test_setup();
 	static void test_cleanup();
-#endif
 	static int start();
 
 	static bool iteration();
