+++ modules/gdscript/gdscript_cache.h
@@ -78,11 +78,9 @@ public:
 	~GDScriptParserRef();
 };
 
-#ifdef TESTS_ENABLED
 namespace GDScriptTests {
 class TestGDScriptCacheAccessor;
 }
-#endif // TESTS_ENABLED
 
 class GDScriptCache {
 	// String key is full path.
@@ -97,9 +95,7 @@ class GDScriptCache {
 	friend class GDScript;
 	friend class GDScriptParserRef;
 	friend class GDScriptInstance;
-#ifdef TESTS_ENABLED
 	friend class GDScriptTests::TestGDScriptCacheAccessor;
-#endif // TESTS_ENABLED
 
 	static GDScriptCache *singleton;
