+++ modules/jsonrpc/SCsub
@@ -8,8 +8,4 @@ env_jsonrpc = env_modules.Clone()
 env_jsonrpc.add_source_files(env.modules_sources, "*.cpp")
 
 if env["tests"]:
-    env_jsonrpc.Append(CPPDEFINES=["TESTS_ENABLED"])
     env_jsonrpc.add_source_files(env.modules_sources, "./tests/*.cpp")
-
-    if env["disable_exceptions"]:
-        env_jsonrpc.Append(CPPDEFINES=["DOCTEST_CONFIG_NO_EXCEPTIONS_BUT_WITH_ALL_ASSERTS"])
