+++ modules/mbedtls/SCsub
@@ -139,11 +139,7 @@ module_obj = []
 env_mbed_tls.add_source_files(module_obj, "*.cpp")
 
 if env["tests"]:
-    env_mbed_tls.Append(CPPDEFINES=["TESTS_ENABLED"])
     env_mbed_tls.add_source_files(module_obj, "./tests/*.cpp")
-
-    if env["disable_exceptions"]:
-        env_mbed_tls.Append(CPPDEFINES=["DOCTEST_CONFIG_NO_EXCEPTIONS_BUT_WITH_ALL_ASSERTS"])
 
 env.modules_sources += module_obj
