+++ modules/zip/SCsub
@@ -10,5 +10,4 @@ env_zip = env_modules.Clone()
 env_zip.add_source_files(env.modules_sources, "*.cpp")
 
 if env["tests"]:
-    env_zip.Append(CPPDEFINES=["TESTS_ENABLED"])
     env_zip.add_source_files(env.modules_sources, "./tests/*.cpp")
