+++ tests/core/io/test_http_client.cpp
@@ -88,7 +88,7 @@ TEST_CASE("[HTTPClient] verify_headers") {
 	ERR_PRINT_ON;
 }
 
-#if defined(MODULE_MBEDTLS_ENABLED) || defined(WEB_EMABLED)
+#if defined(MODULE_MBEDTLS_ENABLED) || defined(WEB_ENABLED)
 TEST_CASE("[HTTPClient] connect_to_host") {
 	Ref<HTTPClient> client = HTTPClient::create();
 	String host = "https://www.example.com";
@@ -99,6 +99,6 @@ TEST_CASE("[HTTPClient] connect_to_host") {
 	Error err = client->connect_to_host(host, port, tls_options);
 	CHECK_MESSAGE(err == OK, "Expected OK for successful connection");
 }
-#endif // defined(MODULE_MBEDTLS_ENABLED) || defined(WEB_EMABLED)
+#endif // defined(MODULE_MBEDTLS_ENABLED) || defined(WEB_ENABLED)
 
 } // namespace TestHTTPClient
