+++ tests/core/io/test_json.cpp
@@ -33,6 +33,7 @@
 TEST_FORCE_LINK(test_json)
 
 #include "core/io/json.h"
+#include "core/variant/typed_array.h"
 
 namespace TestJSON {
