+++ tests/core/io/test_marshalls.cpp
@@ -33,6 +33,7 @@
 TEST_FORCE_LINK(test_marshalls)
 
 #include "core/io/marshalls.h"
+#include "core/object/script_language.h"
 
 namespace TestMarshalls {
