+++ tests/core/object/test_object.cpp
@@ -35,6 +35,7 @@ TEST_FORCE_LINK(test_object)
 #include "core/object/class_db.h"
 #include "core/object/object.h"
 #include "core/object/script_language.h"
+#include "tests/signal_watcher.h"
 
 namespace TestObject {
