+++ tests/core/templates/test_command_queue.cpp
@@ -33,7 +33,6 @@
 TEST_FORCE_LINK(test_command_queue)
 
 #include "core/config/project_settings.h"
-#include "core/math/random_number_generator.h"
 #include "core/object/worker_thread_pool.h"
 #include "core/os/os.h"
 #include "core/os/thread.h"
