+++ tests/display_server_mock.cpp
@@ -34,6 +34,46 @@
 #include "core/input/input_event.h"
 #include "servers/rendering/dummy/rasterizer_dummy.h"
 
+Vector<String> DisplayServerMock::get_rendering_drivers_func() {
+	Vector<String> drivers;
+	drivers.push_back("dummy");
+	return drivers;
+}
+
+void DisplayServerMock::_set_mouse_position(const Point2i &p_position) {
+	if (mouse_position == p_position) {
+		return;
+	}
+	mouse_position = p_position;
+	_set_window_over(Rect2i(Point2i(0, 0), window_get_size()).has_point(p_position));
+}
+void DisplayServerMock::_set_window_over(bool p_over) {
+	if (p_over == window_over) {
+		return;
+	}
+	window_over = p_over;
+	_send_window_event(p_over ? WINDOW_EVENT_MOUSE_ENTER : WINDOW_EVENT_MOUSE_EXIT);
+}
+void DisplayServerMock::_send_window_event(WindowEvent p_event) {
+	if (event_callback.is_valid()) {
+		Variant event = int(p_event);
+		event_callback.call(event);
+	}
+}
+
+bool DisplayServerMock::has_feature(Feature p_feature) const {
+	switch (p_feature) {
+		case FEATURE_MOUSE:
+		case FEATURE_CURSOR_SHAPE:
+		case FEATURE_CLIPBOARD:
+		case FEATURE_CLIPBOARD_PRIMARY:
+			return true;
+		default: {
+		}
+	}
+	return false;
+}
+
 DisplayServer *DisplayServerMock::create_func(const String &p_rendering_driver, DisplayServer::WindowMode p_mode, DisplayServer::VSyncMode p_vsync_mode, uint32_t p_flags, const Vector2i *p_position, const Vector2i &p_resolution, int p_screen, Context p_context, int64_t p_parent_window, Error &r_error) {
 	r_error = OK;
 	RasterizerDummy::make_current();
