+++ tests/display_server_mock.h
@@ -48,50 +48,15 @@ private:
 	String clipboard_text;
 	String primary_clipboard_text;
 
-	static Vector<String> get_rendering_drivers_func() {
-		Vector<String> drivers;
-		drivers.push_back("dummy");
-		return drivers;
-	}
-
+	static Vector<String> get_rendering_drivers_func();
 	static DisplayServer *create_func(const String &p_rendering_driver, DisplayServer::WindowMode p_mode, DisplayServer::VSyncMode p_vsync_mode, uint32_t p_flags, const Vector2i *p_position, const Vector2i &p_resolution, int p_screen, Context p_context, int64_t p_parent_window, Error &r_error);
 
-	void _set_mouse_position(const Point2i &p_position) {
-		if (mouse_position == p_position) {
-			return;
-		}
-		mouse_position = p_position;
-		_set_window_over(Rect2i(Point2i(0, 0), window_get_size()).has_point(p_position));
-	}
-
-	void _set_window_over(bool p_over) {
-		if (p_over == window_over) {
-			return;
-		}
-		window_over = p_over;
-		_send_window_event(p_over ? WINDOW_EVENT_MOUSE_ENTER : WINDOW_EVENT_MOUSE_EXIT);
-	}
-
-	void _send_window_event(WindowEvent p_event) {
-		if (event_callback.is_valid()) {
-			Variant event = int(p_event);
-			event_callback.call(event);
-		}
-	}
+	void _set_mouse_position(const Point2i &p_position);
+	void _set_window_over(bool p_over);
+	void _send_window_event(WindowEvent p_event);
 
 public:
-	bool has_feature(Feature p_feature) const override {
-		switch (p_feature) {
-			case FEATURE_MOUSE:
-			case FEATURE_CURSOR_SHAPE:
-			case FEATURE_CLIPBOARD:
-			case FEATURE_CLIPBOARD_PRIMARY:
-				return true;
-			default: {
-			}
-		}
-		return false;
-	}
+	bool has_feature(Feature p_feature) const override;
 
 	String get_name() const override { return "mock"; }
 
@@ -129,3 +94,109 @@ public:
 		register_create_function("mock", create_func, get_rendering_drivers_func);
 	}
 };
+
+// Utility macros to send an event actions to a given object
+// Requires Message Queue and InputMap to be setup.
+// SEND_GUI_ACTION    - takes an input map key. e.g SEND_GUI_ACTION("ui_text_newline").
+// SEND_GUI_KEY_EVENT - takes a keycode set.   e.g SEND_GUI_KEY_EVENT(Key::A | KeyModifierMask::META).
+// SEND_GUI_KEY_UP_EVENT - takes a keycode set.   e.g SEND_GUI_KEY_UP_EVENT(Key::A | KeyModifierMask::META).
+// SEND_GUI_MOUSE_BUTTON_EVENT - takes a position, mouse button, mouse mask and modifiers e.g SEND_GUI_MOUSE_BUTTON_EVENT(Vector2(50, 50), MOUSE_BUTTON_NONE, MOUSE_BUTTON_NONE, Key::None);
+// SEND_GUI_MOUSE_BUTTON_RELEASED_EVENT - takes a position, mouse button, mouse mask and modifiers e.g SEND_GUI_MOUSE_BUTTON_RELEASED_EVENT(Vector2(50, 50), MOUSE_BUTTON_NONE, MOUSE_BUTTON_NONE, Key::None);
+// SEND_GUI_MOUSE_MOTION_EVENT - takes a position, mouse mask and modifiers e.g SEND_GUI_MOUSE_MOTION_EVENT(Vector2(50, 50), MouseButtonMask::LEFT, KeyModifierMask::META);
+// SEND_GUI_DOUBLE_CLICK - takes a position and modifiers. e.g SEND_GUI_DOUBLE_CLICK(Vector2(50, 50), KeyModifierMask::META);
+
+#define _SEND_DISPLAYSERVER_EVENT(m_event) ((DisplayServerMock *)(DisplayServer::get_singleton()))->simulate_event(m_event);
+
+#define SEND_GUI_ACTION(m_action) \
+	{ \
+		const List<Ref<InputEvent>> *events = InputMap::get_singleton()->action_get_events(m_action); \
+		const List<Ref<InputEvent>>::Element *first_event = events->front(); \
+		Ref<InputEventKey> event = first_event->get()->duplicate(); \
+		event->set_pressed(true); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+#define SEND_GUI_KEY_EVENT(m_input) \
+	{ \
+		Ref<InputEventKey> event = InputEventKey::create_reference(m_input); \
+		event->set_pressed(true); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+#define SEND_GUI_KEY_UP_EVENT(m_input) \
+	{ \
+		Ref<InputEventKey> event = InputEventKey::create_reference(m_input); \
+		event->set_pressed(false); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+#define _UPDATE_EVENT_MODIFIERS(m_event, m_modifiers) \
+	m_event->set_shift_pressed(((m_modifiers) & KeyModifierMask::SHIFT) != Key::NONE); \
+	m_event->set_alt_pressed(((m_modifiers) & KeyModifierMask::ALT) != Key::NONE); \
+	m_event->set_ctrl_pressed(((m_modifiers) & KeyModifierMask::CTRL) != Key::NONE); \
+	m_event->set_meta_pressed(((m_modifiers) & KeyModifierMask::META) != Key::NONE);
+
+#define _CREATE_GUI_MOUSE_EVENT(m_screen_pos, m_input, m_mask, m_modifiers) \
+	Ref<InputEventMouseButton> event; \
+	event.instantiate(); \
+	event->set_position(m_screen_pos); \
+	event->set_button_index(m_input); \
+	event->set_button_mask(m_mask); \
+	event->set_factor(1); \
+	_UPDATE_EVENT_MODIFIERS(event, m_modifiers); \
+	event->set_pressed(true);
+
+#define _CREATE_GUI_TOUCH_EVENT(m_screen_pos, m_pressed, m_double) \
+	Ref<InputEventScreenTouch> event; \
+	event.instantiate(); \
+	event->set_position(m_screen_pos); \
+	event->set_pressed(m_pressed); \
+	event->set_double_tap(m_double);
+
+#define SEND_GUI_MOUSE_BUTTON_EVENT(m_screen_pos, m_input, m_mask, m_modifiers) \
+	{ \
+		_CREATE_GUI_MOUSE_EVENT(m_screen_pos, m_input, m_mask, m_modifiers); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+#define SEND_GUI_MOUSE_BUTTON_RELEASED_EVENT(m_screen_pos, m_input, m_mask, m_modifiers) \
+	{ \
+		_CREATE_GUI_MOUSE_EVENT(m_screen_pos, m_input, m_mask, m_modifiers); \
+		event->set_pressed(false); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+#define SEND_GUI_DOUBLE_CLICK(m_screen_pos, m_modifiers) \
+	{ \
+		_CREATE_GUI_MOUSE_EVENT(m_screen_pos, MouseButton::LEFT, MouseButtonMask::NONE, m_modifiers); \
+		event->set_double_click(true); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
+
+// We toggle _print_error_enabled to prevent display server not supported warnings.
+#define SEND_GUI_MOUSE_MOTION_EVENT(m_screen_pos, m_mask, m_modifiers) \
+	{ \
+		bool errors_enabled = CoreGlobals::print_error_enabled; \
+		CoreGlobals::print_error_enabled = false; \
+		Ref<InputEventMouseMotion> event; \
+		event.instantiate(); \
+		event->set_position(m_screen_pos); \
+		event->set_button_mask(m_mask); \
+		_UPDATE_EVENT_MODIFIERS(event, m_modifiers); \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+		CoreGlobals::print_error_enabled = errors_enabled; \
+	}
+
+#define SEND_GUI_TOUCH_EVENT(m_screen_pos, m_pressed, m_double) \
+	{ \
+		_CREATE_GUI_TOUCH_EVENT(m_screen_pos, m_pressed, m_double) \
+		_SEND_DISPLAYSERVER_EVENT(event); \
+		MessageQueue::get_singleton()->flush(); \
+	}
