+++ tests/scene/test_bit_map.cpp
@@ -32,7 +32,6 @@
 
 TEST_FORCE_LINK(test_bit_map)
 
-#include "core/os/memory.h"
 #include "scene/resources/bit_map.h"
 
 namespace TestBitMap {
