+++ tests/scene/test_code_edit.cpp
@@ -34,7 +34,10 @@ TEST_FORCE_LINK(test_code_edit)
 
 #ifndef ADVANCED_GUI_DISABLED
 
+#include "core/input/input_map.h"
 #include "scene/gui/code_edit.h"
+#include "tests/display_server_mock.h"
+#include "tests/signal_watcher.h"
 
 namespace TestCodeEdit {
