+++ tests/scene/test_graph_node.cpp
@@ -35,7 +35,6 @@ TEST_FORCE_LINK(test_graph_node)
 #ifndef ADVANCED_GUI_DISABLED
 
 #include "scene/gui/graph_node.h"
-#include "scene/main/window.h"
 
 namespace TestGraphNode {
