+++ tests/scene/test_height_map_shape_3d.cpp
@@ -34,9 +34,8 @@ TEST_FORCE_LINK(test_height_map_shape_3d)
 
 #ifndef PHYSICS_3D_DISABLED
 
+#include "core/io/image.h"
 #include "scene/resources/3d/height_map_shape_3d.h"
-#include "scene/resources/image_texture.h"
-#include "tests/test_utils.h"
 
 namespace TestHeightMapShape3D {
