+++ tests/scene/test_navigation_region_2d.cpp
@@ -37,7 +37,6 @@ TEST_FORCE_LINK(test_navigation_region_2d)
 #ifdef MODULE_NAVIGATION_2D_ENABLED
 
 #include "scene/2d/navigation/navigation_region_2d.h"
-#include "scene/main/window.h"
 
 namespace TestNavigationRegion2D {
