+++ tests/scene/test_window.cpp
@@ -32,8 +32,10 @@
 
 TEST_FORCE_LINK(test_window)
 
+#include "core/input/input_map.h" // IWYU pragma: keep // Used by `SEND_GUI_MOUSE_MOTION_EVENT` macro.
 #include "scene/gui/control.h"
 #include "scene/main/window.h"
+#include "tests/display_server_mock.h"
 
 namespace TestWindow {
